requirements.txt
streamlit
opencv-python
numpy
shapely
trimesh
plotly
scipy
pillow
